+++ src/Blizzard/Internal/Html.hs
@@ -1,7 +1,9 @@
+{-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
 
 module Blizzard.Internal.Html
-    ( boolAttribute
+    ( Html
+    , boolAttribute
     , textAttribute
     , documentTag
     , normalTag
@@ -10,9 +12,13 @@ module Blizzard.Internal.Html
 
 
 import Data.Maybe (catMaybes)
+import Data.Monoid (mempty)
 import Data.Text (Text, null)
-import Prelude hiding (null)
-import Text.Blaze.Html
+import Prelude ((>>), ($), (.), Bool(..), Maybe(..), foldl, foldl1)
+import Text.Blaze.Internal ((!), Attribute, AttributeValue, Markup, textValue)
+
+
+type Html = Markup
 
 
 boolAttribute :: Attribute -> Bool -> Maybe Attribute
@@ -25,13 +31,12 @@ textAttribute attr text | null text = Nothing
 textAttribute attr text             = Just . attr . textValue $ text
 
 
-documentTag :: (Html -> Html) -> [Html] -> Html
-documentTag element []       = element $ toHtml ("" :: Text)
-documentTag element children = element $ foldl1 (>>) children
+documentTag :: [Html] -> Html
+documentTag = foldl1 (>>)
 
 
 normalTag :: (Html -> Html) -> [Maybe Attribute] -> [Html] -> Html
-normalTag element attributes []       = foldl (!) element (catMaybes attributes) $ toHtml ("" :: Text)
+normalTag element attributes []       = foldl (!) element (catMaybes attributes) mempty
 normalTag element attributes children = foldl (!) element (catMaybes attributes) $ foldl1 (>>) children